backend/django_service/pyproject.toml
[project]
name = "ansflow-django-service"
version = "0.1.0"
description = "AnsFlow CI/CD Platform - Django Management Service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Django 核心
    "django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "django-environ>=0.10.0",
    # 数据库
    "mysqlclient>=2.1.1",
    "redis>=4.5.0",
    "django-redis>=5.2.0",
    # 认证与权限
    "djangorestframework-simplejwt>=5.2.0",
    "django-oauth-toolkit>=1.7.0",
    "django-guardian>=2.4.0",
    # 消息队列
    "celery>=5.2.0",
    "kombu>=5.2.0",
    "pika>=1.3.0",
    # API 文档
    "drf-spectacular>=0.26.0",
    "drf-spectacular-sidecar>=2023.5.1",
    # 工具库
    "requests>=2.28.0",
    "httpx>=0.24.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    # 监控与日志
    "django-prometheus>=2.2.0",
    "structlog>=22.3.0",
    "sentry-sdk>=1.15.0",
    # 开发工具
    "django-debug-toolbar>=4.0.0",
    "django-extensions>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "factory-boy>=3.2.0",
    "faker>=18.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "django-stubs>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ansflow"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ansflow.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing"
